refactor: Convert the Preset Drive Speed action to structured implementation.
