+++ options.js
@@ -112,14 +112,12 @@ export const WhiteBalanceOption = {
 	},
 }
 
-const PRESET = []
+const PRESET_CHOICES = []
 for (var i = 0; i < 255; ++i) {
 	if (i < 90 || i > 99) {
-		PRESET.push({ id: ('0' + i.toString(16)).slice(-2), label: i })
+		PRESET_CHOICES.push({ id: ('0' + i.toString(16)).slice(-2), label: i })
 	}
 }
-
-export const PRESET_CHOICES = PRESET
 
 export const PresetSaveOption = {
 	id: 'val',
@@ -132,6 +130,49 @@ export const PresetSaveOption = {
 export const PresetRecallOption = {
 	id: 'val',
 	choices: PRESET_CHOICES,
+	choiceToParam: (choice) => {
+		return parseInt(choice, 16)
+	},
+}
+
+export const PresetDriveNumberOption = {
+	id: 'val',
+	choices: PRESET_CHOICES,
+	choiceToParam: (choice) => {
+		return parseInt(choice, 16)
+	},
+}
+
+export const SPEED_CHOICES = [
+	{ id: '18', label: 'Speed 24 (Fast)' },
+	{ id: '17', label: 'Speed 23' },
+	{ id: '16', label: 'Speed 22' },
+	{ id: '15', label: 'Speed 21' },
+	{ id: '14', label: 'Speed 20' },
+	{ id: '13', label: 'Speed 19' },
+	{ id: '12', label: 'Speed 18' },
+	{ id: '11', label: 'Speed 17' },
+	{ id: '10', label: 'Speed 16' },
+	{ id: '0F', label: 'Speed 15' },
+	{ id: '0E', label: 'Speed 14' },
+	{ id: '0D', label: 'Speed 13' },
+	{ id: '0C', label: 'Speed 12' },
+	{ id: '0B', label: 'Speed 11' },
+	{ id: '0A', label: 'Speed 10' },
+	{ id: '09', label: 'Speed 09' },
+	{ id: '08', label: 'Speed 08' },
+	{ id: '07', label: 'Speed 07' },
+	{ id: '06', label: 'Speed 06' },
+	{ id: '05', label: 'Speed 05' },
+	{ id: '04', label: 'Speed 04' },
+	{ id: '03', label: 'Speed 03' },
+	{ id: '02', label: 'Speed 02' },
+	{ id: '01', label: 'Speed 01 (Slow)' },
+]
+
+export const PresetDriveSpeedOption = {
+	id: 'speed',
+	choices: SPEED_CHOICES,
 	choiceToParam: (choice) => {
 		return parseInt(choice, 16)
 	},
